+++ src/main/java/duke/Deadline.java
@@ -1,15 +1,17 @@
+package duke;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Event extends Task {
-    private LocalDateTime from;
-    private LocalDateTime to;
+public class Deadline extends Task {
+//    protected String by;
+    protected LocalDateTime by;
 
-    public Event(String description, String from, String to) {
+    public Deadline(String description, String by) {
         super(description);
-        this.from = parseDateTime(from);
-        this.to = parseDateTime(to);
+        this.by = parseDateTime(by);
+//        this.by = by;
     }
 
     private LocalDateTime parseDateTime(String by) {
@@ -32,12 +34,11 @@ public class Event extends Task {
     @Override
     public String toString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM dd yyyy HHmm");
-        return "[E]" + super.toString() + " (from: " + this.from.format(formatter) + " to: "
-                + this.to.format(formatter) + ")";
+        return "[D]" + super.toString() + " (by: " + this.by.format(formatter) + ")";
     }
 
     @Override
     public String toFile() {
-        return "E | " + super.toFile() + " | " + this.from + " | " + this.to;
+        return "D | " + super.toFile() + " | " + this.by;
     }
 }
